requirements.txt
alembic==1.11.1
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.2
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
inflect==6.0.4
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.7
pymssql==2.2.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
sqlacodegen @ git+https://github.com/bobzomer/sqlacodegen@c9bfe41a7a2e095d1114625ca1fb8e938ba7d81b
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
